+++ 3rdparty/Firebase/macOS/download.sh
@@ -8,5 +8,7 @@ echo "Downloading Firebase"
 
 set -x
 
-wget -nv -O "${LOCATION}/firebase-macOS.tar.bz2" "https://github.com/ntadej/firebase-ios-sdk-binaries/releases/download/v${FIREBASE_VERSION}/firebase-macOS.tar.bz2"
+if [[ ! -f "${LOCATION}/firebase-macOS.tar.bz2" ]]; then
+  wget -nv -O "${LOCATION}/firebase-macOS.tar.bz2" "https://github.com/ntadej/firebase-ios-sdk-binaries/releases/download/v${FIREBASE_VERSION}/firebase-macOS.tar.bz2"
+fi
 tar -C "${LOCATION}" -xf "${LOCATION}/firebase-macOS.tar.bz2"
