+++ 3rdparty/Sparkle/macOS/download.sh
@@ -8,5 +8,7 @@ echo "Downloading Sparkle"
 
 set -x
 
-wget -nv -O "${LOCATION}/Sparkle-${SPARKLE_VERSION}.tar.xz" "https://github.com/sparkle-project/Sparkle/releases/download/${SPARKLE_VERSION}/Sparkle-${SPARKLE_VERSION}.tar.xz"
+if [[ ! -f "${LOCATION}/Sparkle-${SPARKLE_VERSION}.tar.xz" ]]; then
+  wget -nv -O "${LOCATION}/Sparkle-${SPARKLE_VERSION}.tar.xz" "https://github.com/sparkle-project/Sparkle/releases/download/${SPARKLE_VERSION}/Sparkle-${SPARKLE_VERSION}.tar.xz"
+fi
 tar -C "${LOCATION}" -xf "${LOCATION}/Sparkle-${SPARKLE_VERSION}.tar.xz"
